pyproject.toml
[project]
name = "ai-doc-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.1",
    "openai>=2.32.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sentence-transformers>=5.4.1",
    "uvicorn>=0.46.0",
]
